+++ packages/generator/include/combo/mm/save.h
@@ -464,7 +464,9 @@ typedef struct
     u32 hammerGFS:2;
     u32 boomPicto:2;
     u32 bowSlingshot:2;
-    u32 unused:21;
+    u32 stoneGerudoSkull:3;
+    u32 gibdoSpooky:2;
+    u32 unused:16;
 }
 MmExtraItems;
 
@@ -597,6 +599,7 @@ typedef struct ALIGNED(16)
         };
         u8 value;
     } ootSongs;
+    u8 customMask;
 }
 MmCustomSave;
 
@@ -615,5 +618,15 @@ MmCustomSave;
 #define C_SLOT_EQUIP(form, button) (gSaveContext.save.info.itemEquips.cButtonSlots[form][button])
 #define GET_CUR_FORM_BTN_ITEM(btn) ((u8)((btn) == EQUIP_SLOT_B ? BUTTON_ITEM_EQUIP(CUR_FORM, btn) : BUTTON_ITEM_EQUIP(0, btn)))
 #define GET_CUR_FORM_BTN_SLOT(btn) ((u8)((btn) == EQUIP_SLOT_B ? C_SLOT_EQUIP(CUR_FORM, btn) : C_SLOT_EQUIP(0, btn)))
+
+#define GET_WEEKEVENTREG(index) \
+(*(u8*)((u8*)&gSaveContext.save + 0xEF8 + (index)))
+
+#define CHECK_WEEKEVENTREG(flag) \
+(GET_WEEKEVENTREG((flag) >> 8) & ((flag) & 0xFF))
+
+#define WEEKEVENTREG_12_40 ((12 << 8) | 0x40)
+#define WEEKEVENTREG_17_04 ((17 << 8) | 0x04)
+#define WEEKEVENTREG_23_20 ((23 << 8) | 0x20)
 
 #endif /* MM_SAVE_H */