+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add Gerudo Mask, Skull Mask, and Spooky Mask to Majora's Mask.
 - Add hyper enemies/bosses to both games, making them twice as fast.
 - Add Telescope ER.
 - Add Fairy Slingshot to Majora's Mask